Implement the landscape generation window

+++ src/openloco/ui/dropdown.h
@@ -49,19 +49,19 @@ struct format_arg
 
 namespace openloco::ui::dropdown
 {
-    void add(int16_t index, string_id title);
-    void add(int16_t index, string_id title, std::initializer_list<format_arg> l);
-    void add(int16_t index, string_id title, format_arg l);
+    void add(size_t index, string_id title);
+    void add(size_t index, string_id title, std::initializer_list<format_arg> l);
+    void add(size_t index, string_id title, format_arg l);
     int16_t get_highlighted_item();
-    void set_item_disabled(int16_t index);
-    void set_highlighted_item(int16_t index);
-    void set_item_selected(int16_t index);
+    void set_item_disabled(size_t index);
+    void set_highlighted_item(size_t index);
+    void set_item_selected(size_t index);
 
-    void show(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, int8_t count, uint8_t flags);
-    void show_below(window* window, widget_index widgetIndex, int8_t count);
-    void show_below(window* window, widget_index widgetIndex, int8_t count, int8_t height);
-    void show_text(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, int8_t count, uint8_t flags);
-    void show_text_2(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, int8_t count, uint8_t flags);
+    void show(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, size_t count, uint8_t flags);
+    void show_below(window* window, widget_index widgetIndex, size_t count);
+    void show_below(window* window, widget_index widgetIndex, size_t count, int8_t height);
+    void show_text(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, size_t count, uint8_t flags);
+    void show_text_2(int16_t x, int16_t y, int16_t width, int16_t height, colour_t colour, size_t count, uint8_t flags);
 
     void populateCompanySelect(window* window, widget_t* widget);
     company_id_t getCompanyIdFromSelection(int16_t itemIndex);
